+++ config/dispatcher/dispatcher.ex
@@ -135,6 +135,10 @@ defmodule Dispatcher do
     Proxy.forward conn, path, "http://cache/postal-codes/"
   end
 
+  get "/countries/*path", %{ accept: [:json], layer: :api} do
+    Proxy.forward conn, path, "http://cache/countries/"
+  end
+
   # NOTE: resource used
   # NOTE: no request found in frontend
   # match "/groups/*path", %{ accept: [:json], layer: :api } do
@@ -200,6 +204,14 @@ defmodule Dispatcher do
 
   match "/sessions/*path" do
     Proxy.forward conn, path, "http://login/sessions/"
+  end
+
+  ###############################################################
+  # ADDRESS SEARCH
+  ###############################################################
+
+  get "/address-register/*path", %{ accept: [:json], layer: :api} do
+    Proxy.forward conn, path, "http://adressenregister/"
   end
 
   ###############################################################